fastapi_backend/requirements.txt
# FastAPI and Server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database (SQLAlchemy with async support)
sqlalchemy[asyncio]>=2.0.23
aiosqlite>=0.19.0
alembic>=1.13.0
greenlet>=3.0.1

# Pydantic and Settings
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LangChain and AI
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.1.0
langsmith>=0.0.69
openai>=1.3.7

# CORS (Using FastAPI's built-in middleware)
# fastapi-cors>=0.0.1 is not needed 